Remove the legacy embedded connection app and gateway launch path
