+++ packages/mcp-apps/package.json
@@ -8,10 +8,6 @@
       "types": "./dist/skill-created.d.ts",
       "default": "./dist/skill-created.js"
     },
-    "./connection-action": {
-      "types": "./dist/connection-action.d.ts",
-      "default": "./dist/connection-action.js"
-    },
     "./plugin-flow": {
       "types": "./dist/plugin-flow.d.ts",
       "default": "./dist/plugin-flow.js"
